Improved updateScores Function (W.I.P.)
Modified the method for finding 3 and 4 of a kinds to be much more readible and logical. This also makes the check for a full house easier to implement.

Begin working on the small and large straight functions. Currently works in progress.

Changed the IDs of several form elements to be quicker to write.