Implemented logic for Scoring
Created the submitScore function as a method for the PlayFeild object. submitScore uses a swith case to decide what score to give depending on the dice rolled and the score type selected. This function is called by the submit input in the score form. Listed future tasks at the bottom of the function.