Implemented functionality for Local Storage
Added storedData to the PlayFeild Object. This is another object that is meant for quick storage and retrival of data. storedData now contains the turnNum and rollNum as opposed to the PlayFeild object directly. This was done to avoid duplicate data that would always be the same. Added 3 new functions to PlayFeild: canUseLocalStorage checks if the browser supports local storage; store runs the canUseLocalStorage check before turning sotredData into a string so it can be stored inside of local storage; load checks for both local storage avalibility and for stored data before transfering the local storage data into the storedData object.

Added calls to update the data within storedData for each function that used the data. Added a load call after the PlayFeild initialization to ensure page data is loaded in before gameplay starts but after the PlayFeild object and storedData exist.

Fixed a bug that popped up when rolling dice. Either a new storage method or an accidental rewrite caused the dice to not roll on startup. The syntax has been corrected and the dice currently behave as expected. As long as the dice remain initializated at load, no further issues should arrise regarding this bug.