Added a Score class.
This class will manage all the logic related to the score. I did this because at the moment the GameState class still felt to broad and it felt fitting to deconstruct this considering dice are already seperated. The class includes a constructor for score variables, functions for calculatinga nd submitting score, and a function for viewing score types if needed.

Little to no code inside the Score object is currently set. It involves a lot of copying and modifying as do the previous classes. None of the previous classes have been modified to fit in this class yet. All of these are future tasks.