Setup Objects for New Functions.
Cleaned up old code and bad AI code within objects. Changed the StorageHandler class into an object literal, as the current setup only needs a single instance of the object and used static everywhere to get around that. Removed old functions leftover from when Score was not an object. Removed DOM code at top of file now that it is moved to UI object.

Added back in functions that were removed in previous updates as long as they fit the current syntax. I believe that only included the calculateForType function but small tweaks were reinserted as well. The objects are still desperately missing several functions before functionality is restored.