Began reimplementing functions.
Added in the code needed to handle a score including updating current values and passing it to the UI. Removed the getAvalibleScoreTypes function until I need a use for it.

Added score as a parameter in GameState, now when gameState is initialized, it comes with a score.

Preparing to switch all DOM manipulation to JQuery for preference.