Refined UI code and object functions.
After previous class lectures, the project has now switched to jQuery for all DOM related supports. This greatly simplified the code already written and the code that needed to be written for the UIControl. The HTML has also been updated to properly use jQuery.

Most of the functions are now implemented. Gemini offered assistance in organizing classes and calls to make sure everything is running both smoothly and optimally. TODO: re-insert the function for checking which score inputs should be enabled within the UIControl Object.