Went through most major bug fixes.
Implemented the changes necessary to use the highlighted class as opposed to disabling inputs. This included removing !s, reformatting submit to handle 0 score, and so on. Currently still fixing the updateUI (now updateAllUI), as it is getting in the way and is really akward to work with. The HTML and CSS changes are just to add the highlighted class for visual results.