Put project in a most likely working state.
Commented most of the code that needed further explanation. This involved mostly the bigger objects and functions as opposed to the smaller lines. Ended up keeping updateUI as is as it was useful on multiple ocassions to refresh everything. Instead, I moved the enable all input code to the die roll and so on.

Implemented the single coolest trick in this project. I finally found a good use for the CrazeeGame object outside of initialization: changing turns and resetting the game. I passed the CrazeeGame object to the UI it initialized. This allows the UI to make a higher end function call to change the game state, moving gameState's code to somewhere more pratical. It also made a proper reset much easier to implement.