All necessary functional code is complete.
Added the check back to the updateScoreForm function to see if it is the first turn or not. I tried doing this elsewhere, but there were times were I really needed to call updateUI but not update the scoreForm. It now works as expected.

Final bug I found: fixed the HandleSubmitButton function to properly change the state of the button. Turns out jQuery always returns an object, meaning it will never be undefined. However, checking if the length is greater than 0 does the same job as checking for a checked input. Before, this would always enable the submit button and allow for more than 14 turns. Now the game is always 14 turns and is recording as such.

There are features I could add if I have the time, such as extra Crazee scoring options for multiple Crazees, Adding a data tab to view past scores and wipe data, and so on, but I am happy where this game is right now for the project. The first priority is doing the graphics for this project and polishing it up, then additional features can be added.