Began stylizing the page.
Added both Bootstrap styles and JS into the project for quick and solid stylization. Added many classes and divs to the page to apply stylizations.

Transformed the Message Log into a toast element. Updated the JS code to pull the proper elements and use Bootstrap code to make the toast appear and disappear when necessary.

Organized the HTML page for readability. Moved the log Div to the top of the page even if it has a fixed position. Added comments denoting each section of the HTML. Used spacing standards for readability.

Began adding custom styles to the CSS page. Ensured box sizing was always border-box. Forced the page to be at max width to prevent over and under flow. Added a border to the page with the body element. Placed the page in the Game-screen div for improved styles and readability. Added styles for the post log to appear center on smaller screens. Ensured all text was centered by default.

Many more styles, classes, etc. to continue added. Next session should be focused on adding these.