Finished Javascript dice rolling.
Added a function to randomly generate dice images for about a second before showing the final roll. To do this, Dice.roll was updated to only return the roll instead of change the die value. The function rolling all dice was adjusted to account for this, keeping the game functionally intact.

animateRoll was the added function that makes this work. animateRoll includes a callback function to ensure the UI gets updated after the animation. The scoreForm is also now disabled before the roll to prevent submitting a score mid-roll.